+++ app/packages/core/src/components/Modal/Group/DynamicGroup/simple/OrderedDynamicGroup.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import Sample from "../../../Sample";
+import { useGroupContext } from "../../GroupContextProvider";
+import { DynamicGroupCarousel } from "./carousel/DynamicGroupCarousel";
+
+export const OrderedDynamicGroup = () => {
+  const { lookerRefCallback, groupByFieldValue } = useGroupContext();
+
+  if (!groupByFieldValue) {
+    return null;
+  }
+
+  return (
+    <>
+      <DynamicGroupCarousel key={groupByFieldValue} />
+      <Sample lookerRefCallback={lookerRefCallback} />
+    </>
+  );
+};
