+++ app/packages/core/src/components/Modal/Group/DynamicGroup/simple/carousel/DynamicGroupCarousel.tsx
